requirements.txt
apturl==0.5.2
async-generator==1.10
attrs==22.2.0
autopep8==2.0.1
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.7.0
certifi==2022.12.7
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
customtkinter==5.0.4
darkdetect==0.8.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.0
feedparser==6.0.10
gpg===1.13.1-unknown
h11==0.14.0
httplib2==0.14.0
idna==2.8
isc==2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==4.9.2
macaroonbakery==1.3.1
meld==3.20.2
netifaces==0.10.4
notify2==0.3.1
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.10
outcome==1.2.0
pexpect==4.6.0
Pillow==7.0.0
ply==3.11
protobuf==3.6.1
pycairo==1.16.2
pycodestyle==2.10.0
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
PySocks==1.7.1
pyswip==0.2.11
python-apt==2.0.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==4.7.2
sgmllib3k==1.0.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
ssh-import-id==5.10
systemd-python==234
termcolor==2.2.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.14
wadllib==1.3.3
wsproto==1.2.0
xkit==0.0.0
